comments + clean ups
