+++ core/new-gui/src/app/workspace/service/workflow-graph/model/joint-graph-wrapper.ts
@@ -91,13 +91,17 @@ export class JointGraphWrapper {
   private restorePaperOffsetSubject: Subject<Point> = new Subject<Point>();
   // event stream of panning to make mini-map and main workflow paper compatible in offset
   private panPaperOffsetSubject: Subject<Point> = new Subject<Point>();
-
+  // event stream of highlighing a link
   private jointLinkHighlightStream = new Subject<linkIDType>();
+  // event stream of unhighlighing a link
   private jointLinkUnhighlightStream = new Subject<linkIDType>();
+  // event stream of showing the breakpoint button of a link
   private jointLinkBreakpointShowStream = new Subject<linkIDType>();
+  // event stream of hiding the breakpoint button of a link
   private jointLinkBreakpointHideStream = new Subject<linkIDType>();
-
+  // the currently highlighted links' ids
   private currentHighlightedLinks: string[] = [];
+  // the linkIDs of those links with a breakpoint
   private linksWithBreakpoints: string[] = [];
 
   // current zoom ratio
@@ -141,6 +145,7 @@ export class JointGraphWrapper {
     // handle if the current highlighted operator's position is changed,
     // other highlighted operators should move with it.
     this.handleHighlightedOperatorPositionChange();
+    // handle if a link is deleted, it should be unhighlighted
     this.handleLinkDeleteUnhighlight();
   }
 
@@ -226,6 +231,7 @@ export class JointGraphWrapper {
     if (highlightedOperatorIDs.length > 0) {
       this.jointCellHighlightStream.next({ operatorIDs: highlightedOperatorIDs });
     }
+    // if there are highlighed links, unhighlight them
     if (this.currentHighlightedLinks.length > 0) {
       const highlightedLinks = Object.assign([], this.currentHighlightedLinks);
       highlightedLinks.forEach(highlightedLink => this.unhighlightLink(highlightedLink));
@@ -276,27 +282,44 @@ export class JointGraphWrapper {
     return this.jointCellUnhighlightStream.asObservable();
   }
 
+  /**
+   * get the ids of all the links that are currently highlighted
+   */
   public getCurrentHighlightedLinkIDs(): string[] {
     return Object.assign([], this.currentHighlightedLinks);
   }
 
+  /**
+   * get the ids of all the links that have a breakpoint
+   */
   public getLinkIDsWithBreakpoint(): string[] {
     return Object.assign([], this.linksWithBreakpoints);
   }
 
-
+  /**
+   * get the event stream of a link being highlighted.
+   */
   public getLinkHighlightStream(): Observable<linkIDType> {
     return this.jointLinkHighlightStream.asObservable();
   }
 
+  /**
+   * get the event stream of a link being unhighlighted.
+   */
   public getLinkUnhighlightStream(): Observable<linkIDType> {
     return this.jointLinkUnhighlightStream.asObservable();
   }
 
+  /**
+   * get the event stream of showing the breakpoint button of a link
+   */
   public getLinkBreakpointShowStream(): Observable<linkIDType> {
     return this.jointLinkBreakpointShowStream.asObservable();
   }
 
+  /**
+   * get the event stream of hiding the breakpoint button of a link
+   */
   public getLinkBreakpointHideStream(): Observable<linkIDType> {
     return this.jointLinkBreakpointHideStream.asObservable();
   }
@@ -481,6 +504,17 @@ export class JointGraphWrapper {
     element.translate(offsetX, offsetY);
   }
 
+  /**
+   * Highlights the link with given linkID.
+   * Emits an event to the link highlight stream.
+   * If the target link is already highlighted, the action will be ignored.
+   * At current design, there can only be one link highlighted at a time,
+   *  no mutiselect mode for links.
+   * Before a link is highlighted, all the currently highlighted operators will
+   *  be unhighlighted.
+   *
+   * @param linkID
+   */
   public highlightLink(linkID: string): void {
     if (!this.jointGraph.getCell(linkID)) {
       throw new Error(`link with ID ${linkID} doesn't exist`);
@@ -498,16 +532,25 @@ export class JointGraphWrapper {
     this.jointLinkHighlightStream.next({ linkID });
   }
 
+  /**
+   * Unhighlights the given highlighted link.
+   * Emits an event to the link unhighlight stream.
+   * @param unhighlightedLinkID
+   */
   public unhighlightLink(linkID: string): void {
     if (!this.currentHighlightedLinks.includes(linkID)) {
       return;
     }
-
     const unhighlightedLinkIndex = this.currentHighlightedLinks.indexOf(linkID);
     this.currentHighlightedLinks.splice(unhighlightedLinkIndex, 1);
     this.jointLinkUnhighlightStream.next({ linkID });
   }
 
+  /**
+   * Show the breakpoint button of a given link
+   * emits an event to the link breakpoint show stream.
+   * @param linkID
+   */
   public showLinkBreakpoint(linkID: string): void {
     if (!this.linksWithBreakpoints.includes(linkID)) {
       this.linksWithBreakpoints.push(linkID);
@@ -515,6 +558,11 @@ export class JointGraphWrapper {
     this.jointLinkBreakpointShowStream.next({ linkID });
   }
 
+  /**
+   * Hide the breakpoint button of a given link
+   * emits an event to the link breakpoint hide stream.
+   * @param linkID
+   */
   public hideLinkBreakpoint(linkID: string): void {
     if (!this.linksWithBreakpoints.includes(linkID)) {
       return;
@@ -630,6 +678,10 @@ export class JointGraphWrapper {
       });
   }
 
+  /**
+   * Subscribes to link delete event stream,
+   *  unhighlight the removed link if it is currently highlighted
+   */
   private handleLinkDeleteUnhighlight(): void {
     this.getJointLinkCellDeleteStream().subscribe(deletedLinkCell => {
       const deletedLinkID = deletedLinkCell.id.toString();
