+++ core/new-gui/src/app/workspace/service/workflow-graph/model/workflow-action.service.ts
@@ -2,16 +2,14 @@ import { UndoRedoService } from './../../undo-redo/undo-redo.service';
 import { OperatorMetadataService } from './../../operator-metadata/operator-metadata.service';
 import { SyncTexeraModel } from './sync-texera-model';
 import { JointGraphWrapper } from './joint-graph-wrapper';
-import { JointUIService, breakpointButtonSVG } from './../../joint-ui/joint-ui.service';
+import { JointUIService } from './../../joint-ui/joint-ui.service';
 import { WorkflowGraph, WorkflowGraphReadonly } from './workflow-graph';
 import { Injectable } from '@angular/core';
 import { Point, OperatorPredicate, OperatorLink, OperatorPort } from '../../../types/workflow-common.interface';
 
 import * as joint from 'jointjs';
 import { environment } from './../../../../../environments/environment';
-import { isEmpty } from 'rxjs/operators';
 
-type JointPaperEvent = [joint.dia.CellView, JQuery.Event, number, number];
 
 export interface Command {
   execute(): void;
@@ -46,7 +44,6 @@ export class WorkflowActionService {
   private readonly jointGraph: joint.dia.Graph;
   private readonly jointGraphWrapper: JointGraphWrapper;
   private readonly syncTexeraModel: SyncTexeraModel;
-  // private paper: joint.dia.Paper | undefined;
 
   constructor(
     private operatorMetadataService: OperatorMetadataService,
@@ -150,13 +147,6 @@ export class WorkflowActionService {
     paperOptions.model = this.jointGraph;
     return paperOptions;
   }
-
-  /**
-   * setPaper
-   */
-  // public setPaper(paper: joint.dia.Paper) {
-  //   this.paper = paper;
-  // }
 
   /**
    * Adds an opreator to the workflow graph at a point.
@@ -354,6 +344,9 @@ export class WorkflowActionService {
     this.executeAndStoreCommand(command);
   }
 
+  /**
+   * set a given link's breakpoint properties to specific values
+   */
   public setLinkBreakpoint(linkID: string, newBreakpoint: object): void {
     const prevBreakpoint = this.getTexeraGraph().getLinkWithID(linkID).breakpointProperties;
     const command: Command = {
@@ -377,6 +370,11 @@ export class WorkflowActionService {
     this.executeAndStoreCommand(command);
   }
 
+  /**
+   * Set the link's breakpoint property to empty to remove the breakpoint
+   *
+   * @param linkID
+   */
   public removeLinkBreakpoint(linkID: string): void {
     this.setLinkBreakpoint(linkID, {});
   }
@@ -429,55 +427,6 @@ export class WorkflowActionService {
     const jointLinkCell = JointUIService.getJointLinkCell(link);
     this.jointGraph.addCell(jointLinkCell);
     // JointJS link add event will propagate and trigger Texera link add
-
-    // const InfoButton = joint.linkTools.Button.extend({
-    //   name: 'info-button',
-    //   options: {
-    //       markup: [{
-    //           tagName: 'circle',
-    //           selector: 'button',
-    //           attributes: {
-    //               'r': 7,
-    //               'fill': '#001DFF',
-    //               'cursor': 'pointer',
-    //               'event': 'breakpoint',
-    //           },
-    //       }, {
-    //           tagName: 'path',
-    //           selector: 'icon',
-    //           attributes: {
-    //               'd': 'M -2 4 2 4 M 0 3 0 0 M -2 -1 1 -1 M -1 -4 1 -4',
-    //               'fill': 'none',
-    //               'stroke': '#FFFFFF',
-    //               'stroke-width': 2,
-    //               'pointer-events': 'none'
-    //           }
-    //       }],
-    //       event: 'breakpoint',
-    //       distance: 60,
-    //       offset: 0,
-    //       id: jointLinkCell.id
-    //       // action: function(evt: JQuery.Event) {
-    //       //   if (evt.toElement) {
-    //       //     console.log('what can I do?!!!!');
-    //       //     // evt.toElement.dispatchEvent(new Event('tool:remove'));
-    //       //   }
-    //       // }
-    //         // console.log(event);
-    //         // console.log(this.id);
-    //         // WorkflowActionService.deleteLinkWithID(this.id);
-    //   }
-    // });
-
-    // if (this.jointGraph) {
-    //   const linkview = jointLinkCell.findView(this.paper);
-    //   const removeButton = new joint.linkTools.Remove();
-    //   const infoButton = new InfoButton();
-    //   const toolsView = new joint.dia.ToolsView({
-    //     tools: [removeButton, infoButton]
-    //   });
-    //   linkview.addTools(toolsView);
-    // }
   }
 
   private deleteLinkWithIDInternal(linkID: string): void {
@@ -495,6 +444,7 @@ export class WorkflowActionService {
     this.texeraGraph.setOperatorAdvanceStatus(operatorID, newShowAdvancedStatus);
   }
 
+  // modifies the link breakpoint property
   private setLinkBreakpointInternal(linkID: string, breakpoint: object) {
     this.texeraGraph.setLinkBreakpoint(linkID, breakpoint);
   }
